// src/types/Product.d.ts
export interface Products {
  title: string;
  price: number;
  description: string;
  hex: string;
  gender: string;
  color: string;
  image: string[];
  tags: string[];
  category: string;
  _id: string;
  preview: string;
  size: string[];
  quantity: number;
  selectedSize?: string | undefined;
  _createdAt: string;
}